Add files via upload
🖼️ CIFAR-10 Dataset (Custom CNN)

In this project, I experimented with the CIFAR-10 dataset using a custom CNN model. The architecture design was restricted, which led to relatively low accuracy compared to state-of-the-art models. Additionally, I trained the model on a limited subset (5,000 training images and 10,000 test images) to test model performance under constrained conditions.